+++ pypeit/scripts/run_pypeit.py
@@ -86,9 +86,9 @@ def main(args):
     # Load PypeIt file to get the spectrograph (might happen twice but that is ok)
     pypeitSetup = pypeitsetup.PypeItSetup.from_pypeit_file(args.pypeit_file)
 
-    pypeIt = pypeit.instantiate_me(pypeitSetup.spectrograph, args.pypeit_file,
-                                   verbosity=args.verbosity,reuse_masters=args.use_masters,
-                                   overwrite=args.overwrite, logname=logname, show=args.show)
+    pypeIt = pypeit.PypeIt(pypeitSetup.spectrograph, args.pypeit_file,
+                           verbosity=args.verbosity,reuse_masters=args.use_masters,
+                           overwrite=args.overwrite, logname=logname, show=args.show)
 
     # JFH I don't see why this is an optional argument here. We could allow the user to modify an infinite number of parameters
     # from the command line? Why do we have the PypeIt file then? This detector can be set in the pypeit file.
